Improve history deletion when saving crawled data.
There was a bug when the user has many entries, and MAL limits the history to 300 entries. So the code was always deleting entries with less than 21 days, but this is not still the case. Sometimes, MAL can hide older entries because the user has reached this limit, so it causes an inappropriate deletion of entries that can not be seen on the user history page.

Now we will delete based on the oldest entry. We want to make sure that we are not losing any entry by a deletion, so we delete just the entries that have been already imported but are in theory present on the history payload crawled from MAL. Also, we are going to limit the deletion to 300 records, so it keeps concise with the MAL behavior.

Special thanks to MAL's user YoshePlays who has reported this bug.