Add feature flag count each entry as an activity.
Some users are requesting to change the default way of calculating activities. They prefer to count each line entry as an activity on that day despite the amount that has been recorded. I'm creating a feature flag toggle this behavior as needed, so the other users can still use them without noticeable changes.

Special thanks to the YoshePlays user who pointed that out