Fix issue when converting entries timestamps to UTC
When converting timestamps from the user timezone to UTC the offset was being set wrongly due some weird conversion. Now we are going to parse the natural timestamp as a pure Time object and then convert it to UTC time based on the current user timezone time

Big thanks to @Squashbucklr who noticed the bug and reported it!

Fix issue when converting entries timestamps to UTC

When converting timestamps from the user timezone to UTC the offset was being set wrongly due some weird conversion. Now we are going to parse the natural timestamp as a pure Time object and then convert it to UTC time based on the current user timezone time

Big thanks to @Squashbucklr who noticed the bug and reported it!